Begin changing to a unified split-list for all datasets
This will remove the need to know the number of rows ahead of time.
Further, It will ensure that if the splitted frame is not in the
environment, that drake won't freak out about it. Also, TODO: include a
`character_only` argument to pass a character name, rather than an
actual object name.