Add light parallelism to hash_list() and loadd()
See lightly_parallelize(). Should help mitigate the
bottlenecks exposed by @dapperjapper, as long as
the user is not on Windows.

I chose mclapply because it can't hurt. Stuff only gets
faster. Even parLapply is way to high-overhead for the
usual workflow with lots of tiny functions and files to hash
over multiple stages before any targets are even built.