Deprecate plan() and workflow() in favor of workplan()
I have been struggling with a new name for this one.
I used plan() originally, but there is a glaring conflict
with future::plan(). Then I thought workflow() would be
more appropriate, but the first argument to make() is still
called plan. Finally, I thought of workplan() and decided that
it was a more appropriate replacement. I will remove deprecated
functoins plan() and workflow() about a year from now.