Write a `.gitignore` file **inside** the `.drake` cache.
`.drake/.gitignore` just contans "*" to
hide everything in the cache from `git`.
The `.gitignore` file is only created when the cache is first created,
so users can remove it and it will stay removed in subsequent `make()`s.
Only works with the default cache, so users with custom caches
are on their own.

cc @uribo @krlmlr