Standardize away `ignore()` using proper code analysis
Now walking the AST. `ignore(xyz + blablabla + fun(x))` turns
into `ignore()`, and the same happens with `drake::ignore(...)`
and `drake:::ignore(...)`. Using a for loop, which works if
we're careful and is not likely to bottleneck here.
Leaving an `ignore()` behind makes sure the code can
still be parsed (e.g. `x / ignore(7)`).

cc @krlmlr